set-version handles missing doc

+++ build-tools/set-version.sh
@@ -16,10 +16,16 @@ old_version=$(cat $DIR/../VERSION)
 echo Updating version from \"$old_version\" to \"$version\"
 echo "    updating $DIR/../VERSION"
 echo $version > $DIR/../VERSION
-echo "    updating $DIR/../doc/conf.py"
-sed -i.tmp "s/version = u'.*'/version = u'$version'/" $DIR/../doc/conf.py
-sed -i.tmp "s/release = u'.*'/release = u'$version'/" $DIR/../doc/conf.py
-rm $DIR/../doc/conf.py.tmp
+
+if [ -f $DIR/../doc/conf.py ]; then
+  echo "    updating $DIR/../doc/conf.py"
+  sed -i.tmp "s/version = u'.*'/version = u'$version'/" $DIR/../doc/conf.py
+  sed -i.tmp "s/release = u'.*'/release = u'$version'/" $DIR/../doc/conf.py
+  rm $DIR/../doc/conf.py.tmp
+else
+  echo "    skipping $DIR/../doc/conf.py ... does not exist"
+fi
+
 
 for package_dir in $DIR/../utils/loomengine_utils $DIR/../worker/loomengine_worker $DIR/../server/loomengine_server $DIR/../client/loomengine
